+++ src/main/java/roomescape/service/LoginService.java
@@ -4,9 +4,9 @@ import org.springframework.stereotype.Service;
 import roomescape.domain.Member;
 import roomescape.domain.MemberRepository;
 import roomescape.domain.MemberRole;
-import roomescape.exception.auth.InvalidTokenException;
-import roomescape.exception.auth.UnauthorizedEmailException;
-import roomescape.exception.auth.UnauthorizedPasswordException;
+import roomescape.exception.login.InvalidTokenException;
+import roomescape.exception.login.UnauthorizedEmailException;
+import roomescape.exception.login.UnauthorizedPasswordException;
 import roomescape.service.dto.LoginCheckResponse;
 import roomescape.service.dto.LoginRequest;
 import roomescape.service.dto.SignupRequest;
